add npm history package add getQuery and setQuery functions to helpers.js file
