+++ client/package.json
@@ -8,6 +8,7 @@
   "dependencies": {
     "axios": "^0.18.0",
     "date-fns": "^1.30.1",
+    "history": "^4.10.1",
     "jsonwebtoken": "^8.5.0",
     "moment": "^2.24.0",
     "node-sass": "^4.11.0",
@@ -21,7 +22,7 @@
     "react-dom": "^16.13.1",
     "react-quill": "^1.3.3",
     "react-redux": "^6.0.1",
-    "react-router-dom": "^5.0.0",
+    "react-router-dom": "^5.1.2",
     "react-swipe": "^6.0.4",
     "redux": "^4.0.1",
     "redux-devtools-extension": "^2.13.8",
